sailui/topnav: create component
